Add support for ... syntax for recursive analysis
Refactor to enable using the ./... syntax to recursively process a
directory. This changes the behaviour of the program as before the
analysis was always recursive unless the "-n" was used. Because the
analysis is now only recursive when a path ends in ... the "-n" flag is
no longer necessary.

So this changes the program in two breaking ways:

- "-n" is no longer useful.
- Old paths that would be analyzed recursively will no longer be
   analyzed recursively.